pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Timur", email = "bas9339@gmail.com"},
]
dependencies = [
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.6",
    "spotipy>=2.23.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

